created inat model, cant connect no username password
