+++ opentreemap/opentreemap/settings/default_settings.py
@@ -294,6 +294,7 @@ INSTALLED_APPS = (
     'django.contrib.postgres',
     'django_js_reverse',
     'webpack_loader',
+    'background_task'
 )
 
 I18N_APPS = (
