+++ requirements.txt
@@ -11,6 +11,7 @@ chardet==3.0.4
 # https://docs.djangoproject.com/en/1.10/releases/#id2
 Django==1.11.16  # rq.filter: >=1.11,<1.12
 django-apptemplates==1.3
+django-background-tasks==1.2.5
 django-contrib-comments==1.8.0
 django-js-reverse==0.7.3
 django-queryset-csv==1.0.2             # https://github.com/azavea/django-queryset-csv/commits/master
